Fixes issues with the previous merge
