+++ lib/src/rule_creator_component.html
@@ -96,6 +96,7 @@
                                   [minlength]="1"
                                   pattern="\d+"
                                   patternMessage="Field should only contain an advertiser ID">
+                        </bs-input>
                     </div>
                 </div>
                 <div *ngSwitchCase="'Change line item bidding strategy'">
