+++ server/lib/service/dv360.dart
@@ -7,6 +7,7 @@ import 'package:http/http.dart';
 import '../model/action.dart';
 import '../model/change_line_item_bidding_strategy_action.dart';
 import '../model/change_line_item_status_action.dart';
+import '../model/duplicate_line_item_action.dart';
 import '../model/rule.dart';
 import '../model/scope.dart';
 import '../proto/rule.pb.dart' as proto;
@@ -70,6 +71,7 @@ class DisplayVideo360Client {
     await changeLineItemStatus(Int64.parseInt(duplicate.advertiserId),
         Int64.parseInt(duplicate.lineItemId), currentStatus);
   }
+
   /// Changes the bidding strategy of the line item to [strategy].
   ///
   /// Throws an [ApiRequestError] if API returns an error.
@@ -151,7 +153,7 @@ class DisplayVideo360Client {
     final changeStatusAction = action as ChangeLineItemStatusAction;
 
     final shortStatusName = proto.ChangeLineItemStatusParams_Status.valueOf(
-        changeStatusAction.statusValue);
+        changeStatusAction.statusValue.index);
     final status = 'ENTITY_STATUS_${shortStatusName.name}';
 
     await changeLineItemStatus(
@@ -188,9 +190,8 @@ class DisplayVideo360Client {
         strategy = 'PERFORMANCE_GOAL';
         break;
       default:
-        throw UnsupportedError(
-            '${changeStrategyAction.biddingStrategy} is an '
-                'invalid bidding strategy.');
+        throw UnsupportedError('${changeStrategyAction.biddingStrategy} is an '
+            'invalid bidding strategy.');
     }
 
     if (changeStrategyAction.biddingStrategy != BidStrategyType.fixed) {
@@ -216,7 +217,7 @@ class DisplayVideo360Client {
         default:
           throw UnsupportedError(
               '${changeStrategyAction.performanceGoal} is an '
-                  'invalid performance goal.');
+              'invalid performance goal.');
       }
     }
